ERROR FIXED

+++ FutureTools_LITE.R
@@ -3,6 +3,12 @@ rm(list=ls())
 library("beepr")
 setwd("C:/Temp/")
 
+##
+msg.path <- "C:/Temp/"
+price.path <- "C:/Temp/msg/"
+realdata.path <- "C:/Users/linus/Documents/Project/9.Shared.Data/8.forSmartAPI/"
+
+##
 #### 設定額外函式位置 #### 
 source("C:/Users/linus/Documents/Project/6.APITols/Order_module_base.R")
 source("C:/Users/linus/Documents/Project/6.APITols/Order_module_custom.R")
@@ -10,91 +16,74 @@ source("C:/Users/linus/Documents/Project/6.APITols/Order_module_POSITION.R")
 source("C:/Users/linus/Documents/Project/6.APITols/Order_module_SIMUServer.R")
 source("C:/Users/linus/Documents/Project/6.APITols/Order_module_AGENTServer.R")
 source("C:/Users/linus/Documents/Project/6.APITols/m_libs.R")
-#主程式設定
-source("C:/Users/linus/Documents/Project/6.APITols/FutureTools_config.R")
+##
+switch.check.if.deal <-TRUE
+transaction <-NULL #交易結果訊息向量
+MXFSIMU.Name <- "MXFSIMU"
+MXFSIMU.file <- filename.gen(name=MXFSIMU.Name)
+MXFSIMU.data.path <- paste0(msg.path, "/", MXFSIMU.Name, "/_Match.txt")
+MXFSIMU.forSERVER.filename <- paste0(msg.path, filename.gen(x="log"))
 
+##
+Product <-"MXFK1"
+Price <-0
+BorS <- "" #買(B)或賣(S)
+Daytrade <-"1" #設定當沖(否1是0)
+switch.stopPORT <-5
+.path <-extra.data(name="switch_to.ma", p.mode = "path")
+append.to.file(data=switch.stopPORT
+               , path=.path)
 
-# 
-# ##
-# switch.check.if.deal <-FALSE
-# transaction <-NULL #交易結果訊息向量
-# MXFSIMU.Name <- "MXFSIMU"
-# MXFSIMU.file <- filename.gen(name=MXFSIMU.Name)
-# # MXFSIMU.Server <- FALSE
-# MXFSIMU.data.path <- paste0(msg.path, "/", MXFSIMU.Name, "/_Match.txt")
-# MXFSIMU.forSERVER.filename <- paste0(msg.path, filename.gen(x="log"))
-# 
-# # MXFSIMU.source.data.path <- paste0(msg.path, "/", MXFSIMU.Name, "/_Match_source.txt")
-# # MXFSIMU.source.data.path <-finacial.dataparg.gen(realdata.path, .input, Product, Product.file)
-# 
-# ##
-# Product <-"MXFJ1"
-# Price <-0
-# BorS <- "" #買(B)或賣(S)
-# Daytrade <-"1" #設定當沖(否1是0)
-# switch.stopPORT <-5
-# .path <-extra.data(name="switch_to.ma", p.mode = "path")
-# append.to.file(data=switch.stopPORT
-#                , path=.path)
-# 
-# DateFolder <- ""
-# result <- "  "
-# Qty <-1
-# gear <-0
-# BASE_portfolio <- 3  #無虧損停利價差
-# Stop_portfolio <- 10 #動態停利價差
-# default.enable_stopPORTFOLIO <- 15 #固定停利價差
-# Max.DDM <- 0
-# default.PORTFOLIO.buffer <-5
-# Keep.NOLOSS.ratio <-2
-# # Stop_loss <- 0 #停損
-# PCL <- 0 #多空代號 1 -1
-# Price.buyin <- 0
-# simu <-TRUE
-# Auto.positionCLOSE <-FALSE
-# enable.STABLE.Stop.PORT <-TRUE #停利功能>>預設非動態停利
-# enable.defaultPORT.check <-TRUE #開啟停利功能
-# 
-# TRENDMark.LONG <-FALSE
-# TRENDMark.SHORT <-FALSE
-# 
-# preWORK.name <-c("CURRENTBAR.ADDED")
-# preWORK.check <-c(rep(FALSE, length(preWORK.name)))
-# 
-# safe.Close <- TRUE #TRUE表使用緊急平倉來平倉
-# Stop_portfolio.type <-c("(1)MDD", "(2)RsiOVER_SB", "(3)Bolling")
-# Stop_portfolio.code <-1
-# # names(Stop_portfolio.type) <- c("7", "77")
-# Stop_loss.type <-c("(1)RsiREVERSAL", "(2)ResearchLINE", "(3)ExtremeLINE", "(4)Bolling", "(5)PolarSTAR")
-# Stop_loss.code <-1
-# next.step <- ""
-# Price.reachLIMITED.times.Limited <-2
-# 
-# get.hour <- as.numeric(format(Sys.time(), "%H"))
-# get.sysDate <-  Sys.Date()
-# if (get.hour <8){get.sysDate = get.sysDate -1 } 
-# date.format <- gsub("-", "", get.sysDate)
-# #date.format <- gsub("-", "", Sys.Date())
-# 
-# # Product.file <- paste0(date.format, "_Match.txt")
-# Product.file <- filename.gen(name=date.format)
-# SECURTIES.data.path <-finacial.dataparg.gen(realdata.path, date.format, Product, Product.file)
-# 
-# #設定預設資料源<證卷商>
-# switch.DATA.Source <-TRUE #T表示證卷商
-# data.path <- data.source.switch(switch.DATA.Source)
-# # data.path <- paste0("C:/Users/linus/Documents/Project/9.Shared.Data/8.forSmartAPI/"
-# #                     , date.format, "/"
-# #                     , Product, "/"
-# #                     , Product.file)
-# 
-# enable.STABLE.Stop.PORT.path  <- extra.data(name="enable.STABLE.Stop.PORT", p.mode = "path") #default固定停利
-# enable.onlyMDD.path  <- extra.data(name="enable.onlyMDD", p.mode = "path") #MDD停利
-# enable.RSI.TrendADDED.path  <- extra.data(name="enable.RSI.TrendADDED", p.mode = "path") #RSI超買超賣停利
-# enable.Bolling.path  <- extra.data(name="enable.BollingPATH.ADDED", p.mode = "path") #布林通道停利
-# DMSS.path  <- extra.data(name="DMSS", p.mode = "path") #停止虛擬資料伺服器
-# DAGS.path  <- extra.data(name="DAGS", p.mode = "path") #停止代理人伺服器
-# RAGS.path  <- extra.data(name="RESET_AGENT.SERVERE", p.mode = "path") #重設代理人伺服器
+DateFolder <- ""
+result <- "  "
+Qty <-1
+gear <-0
+BASE_portfolio <- 3  #無虧損停利價差
+Stop_portfolio <- 10 #動態停利價差
+default.enable_stopPORTFOLIO <- 15 #固定停利價差
+Max.DDM <- 0
+default.PORTFOLIO.buffer <-5
+Keep.NOLOSS.ratio <-2
+PCL <- 0 #多空代號 1 -1
+Price.buyin <- 0
+simu <-TRUE
+Auto.positionCLOSE <-FALSE
+enable.STABLE.Stop.PORT <-TRUE #停利功能>>預設非動態停利
+enable.defaultPORT.check <-TRUE #開啟停利功能
+
+TRENDMark.LONG <-FALSE
+TRENDMark.SHORT <-FALSE
+
+preWORK.name <-c("CURRENTBAR.ADDED")
+preWORK.check <-c(rep(FALSE, length(preWORK.name)))
+
+safe.Close <- TRUE #TRUE表使用緊急平倉來平倉
+Stop_portfolio.type <-c("(1)MDD", "(2)RsiOVER_SB", "(3)Bolling")
+Stop_portfolio.code <-1
+Stop_loss.type <-c("(1)RsiREVERSAL", "(2)ResearchLINE", "(3)ExtremeLINE", "(4)Bolling", "(5)PolarSTAR")
+Stop_loss.code <-1
+next.step <- ""
+Price.reachLIMITED.times.Limited <-2
+
+get.hour <- as.numeric(format(Sys.time(), "%H"))
+get.sysDate <-  Sys.Date()
+if (get.hour <8){get.sysDate = get.sysDate -1 } 
+date.format <- gsub("-", "", get.sysDate)
+
+Product.file <- filename.gen(name=date.format)
+SECURTIES.data.path <-finacial.dataparg.gen(realdata.path, date.format, Product, Product.file)
+
+#設定預設資料源<證卷商>
+switch.DATA.Source <-TRUE #T表示證卷商
+data.path <- data.source.switch(switch.DATA.Source)
+
+enable.STABLE.Stop.PORT.path  <- extra.data(name="enable.STABLE.Stop.PORT", p.mode = "path") #default固定停利
+enable.onlyMDD.path  <- extra.data(name="enable.onlyMDD", p.mode = "path") #MDD停利
+enable.RSI.TrendADDED.path  <- extra.data(name="enable.RSI.TrendADDED", p.mode = "path") #RSI超買超賣停利
+enable.Bolling.path  <- extra.data(name="enable.BollingPATH.ADDED", p.mode = "path") #布林通道停利
+DMSS.path  <- extra.data(name="DMSS", p.mode = "path") #停止虛擬資料伺服器
+DAGS.path  <- extra.data(name="DAGS", p.mode = "path") #停止代理人伺服器
+RAGS.path  <- extra.data(name="RESET_AGENT.SERVERE", p.mode = "path") #重設代理人伺服器
 
 #MAIN# 
 
@@ -193,7 +182,9 @@ repeat
             CA ={result <- CancelAll()},
             "5"  ={
                     result <- CancelAll()
-                    transaction <-account.info(code=result)
+                    # transaction <-account.info(code=result)
+                    print(paste("回傳結果(取消下單) :", result))
+                    
             },
             QA ={result <- QueryAllOrder()},
             QO ={result <- QueryOnOpen()},
@@ -207,7 +198,8 @@ repeat
                     msg.file  <- extra.data(name="close.ALLPOSITION", p.mode = "path") 
                     file.create(msg.file)
                     result <- ClosePositionAll()
-                    transaction <-account.info(code=result)
+                    # transaction <-account.info(code=result)
+                    print(paste("回傳結果(全平倉) :", result))
                     
                   },
             "4"  ={
@@ -216,11 +208,11 @@ repeat
                     Price <- Price.current()
                     result <- Place.OrderLMT()
                     # result <- transaction.MGR(pdt=Product, bors = "B")
-                    transaction <-account.info(code=result)
-                    print(paste("回傳結果 :", result, ">>", transaction))
+                    transaction <-account.info(code=result) #依下單回傳訊息解碼成文字向量
+                    print(paste("回傳結果 :", result, ">>", transaction[1]))
                     # Price.buyin <- as.numeric(Price)
                     Price.buyin <- as.numeric(account.info(info =transaction
-                                                           , name = "price" ))
+                                                           , p.mode = "by.name" , name = "price" ))
 
                     PCL <- 1
                     # transaction <-transaction.all(bs="B")
@@ -230,9 +222,9 @@ repeat
                     # 
                     # PCL <- 1
                     # 
-                    # if(check.if.deal())
-                    # {
-                      # print(paste("[下單觸發] 完全成交"))
+                    if(check.if.deal(force = TRUE, decoded.info =transaction ))
+                    {
+                      print(paste("[訊息] 執行成交後續設定"))
                       .path <- extra.data(name="price.Buyin", p.mode = "path")
                       .PCL.path <- extra.data(name="price.PCL", p.mode = "path")
                       .msg.path <- extra.data(name="create.positionLONG", p.mode = "path")
@@ -246,7 +238,7 @@ repeat
                       {
                         next.step <- "7"
                       }                     
-                    # }
+                    }
        
                   },
             "2"  ={
@@ -255,10 +247,11 @@ repeat
                     Price <- Price.current()
                     result <- Place.OrderMKT()
                     transaction <-account.info(code=result)
-
+                    print(paste("回傳結果 :", result, ">>", transaction[1]))
                     # Price.buyin <- as.numeric(Price)
                     Price.buyin <- as.numeric(account.info(info =transaction
-                                                           , name = "price" ))
+                                                           , p.mode = "by.name" , name = "price" ))
+                    
                     PCL <- 1
                     # transaction <-transaction.all(bs="B")
                     # 
@@ -267,8 +260,9 @@ repeat
                     # 
                     # PCL <- 1
                     # 
-                    # if(check.if.deal())
-                    # {
+                    if(check.if.deal(force = TRUE, decoded.info =transaction ))
+                    {
+                      print(paste("[訊息] 執行成交後續設定"))
                       .path <- extra.data(name="price.Buyin", p.mode = "path")
                       .PCL.path <- extra.data(name="price.PCL", p.mode = "path")
                       .msg.path <- extra.data(name="create.positionLONG", p.mode = "path")
@@ -282,7 +276,7 @@ repeat
                       {
                         next.step <- "7"
                       }                     
-                    # }              
+                    }              
                   },
             "6"  ={
                     #Qty <-1
@@ -291,37 +285,36 @@ repeat
                     result <- Place.OrderLMT()
                     # result <- transaction.MGR(bors = "S")
                     transaction <-account.info(code=result)
-                    print(paste("回傳結果 :", result, ">>", transaction))
-
+                    print(paste("回傳結果 :", result, ">>", transaction[1]))
                     # Price.buyin <- as.numeric(Price)
                     Price.buyin <- as.numeric(account.info(info =transaction
-                                                           , name = "price" ))
+                                                           , p.mode = "by.name" , name = "price" ))
+                    
                     PCL <- -1
-                    # transaction <-transaction.all(bs="S")
+                    # transaction <-transaction.all(bs="B")
                     # 
                     # Price.buyin <- as.numeric(account.info(p.mode ="by.name", info =transaction
                     #                                        , name = "price" ))
                     # 
                     # PCL <- 1
                     # 
-                    # if(check.if.deal())
-                    # {
+                    if(check.if.deal(force = TRUE, decoded.info =transaction ))
+                    {
+                      print(paste("[訊息] 執行成交後續設定"))
                       .path <- extra.data(name="price.Buyin", p.mode = "path")
                       .PCL.path <- extra.data(name="price.PCL", p.mode = "path")
-                      .msg.path <- extra.data(name="create.positionSHORT", p.mode = "path")
-                      
+                      .msg.path <- extra.data(name="create.positionLONG", p.mode = "path")
                       unlink(.path)
                       append.to.file(data=Price.buyin
                                      , path=.path)
                       append.to.file(data=PCL
                                      , path=.PCL.path)
                       file.create(.msg.path)
-                      
                       if(Auto.positionCLOSE)
                       {
                         next.step <- "7"
-                      }
-                    # }
+                      }                     
+                    }
                  },
             "8"  ={
                     #Qty <-1
@@ -329,45 +322,45 @@ repeat
                     Price <- Price.current()
                     result <- Place.OrderMKT()
                     transaction <-account.info(code=result)
-
+                    print(paste("回傳結果 :", result, ">>", transaction[1]))
                     # Price.buyin <- as.numeric(Price)
                     Price.buyin <- as.numeric(account.info(info =transaction
-                                                           , name = "price" ))
+                                                           , p.mode = "by.name" , name = "price" ))
+                    
                     PCL <- -1
-                    # transaction <-transaction.all(bs="S")
+                    # transaction <-transaction.all(bs="B")
                     # 
                     # Price.buyin <- as.numeric(account.info(p.mode ="by.name", info =transaction
                     #                                        , name = "price" ))
                     # 
                     # PCL <- 1
                     # 
-                    # if(check.if.deal())
-                    # {
+                    if(check.if.deal(force = TRUE, decoded.info =transaction ))
+                    {
+                      print(paste("[訊息] 執行成交後續設定"))
                       .path <- extra.data(name="price.Buyin", p.mode = "path")
                       .PCL.path <- extra.data(name="price.PCL", p.mode = "path")
-                      .msg.path <- extra.data(name="create.positionSHORT", p.mode = "path")
-                      
+                      .msg.path <- extra.data(name="create.positionLONG", p.mode = "path")
                       unlink(.path)
                       append.to.file(data=Price.buyin
                                      , path=.path)
                       append.to.file(data=PCL
                                      , path=.PCL.path)
                       file.create(.msg.path)
-                      
                       if(Auto.positionCLOSE)
                       {
                         next.step <- "7"
-                      }
-                    # }             
+                      }                     
+                    }             
                 },
-            # # 多重建倉法
-            "9" ={
-                    Position.multi.create()
-                    if(Auto.positionCLOSE)
-                    {
-                      next.step <- "7"
-                    }
-            },
+            # # # 多重建倉法
+            # "9" ={
+            #         Position.multi.create()
+            #         if(Auto.positionCLOSE)
+            #         {
+            #           next.step <- "7"
+            #         }
+            # },
             # # 切換保本模式
             # "3" ={
             #       price.diff <-abs(as.numeric(readline("KEEP.CASH MOD(adjust buffer) :")))
